requirements/local.txt
-r base.txt

psycopg2-binary==2.9.9
flake8==6.0.0
black==23.1.0
isort==5.12.0
watchfiles==0.18.1
django-colorfield==0.9.0
django-allauth==0.52.0
djangorestframework-simplejwt==5.2.2
dj-rest-auth==3.0.0
django-modeltranslation==0.18.10
